+++ packages/core/navigation-next/src/__tests__/unit/server-side.js
@@ -1,0 +1,17 @@
+/**
+ * @jest-environment node
+ */
+// @flow
+
+import { getExamplesFor } from '@atlaskit/build-utils/getExamples';
+import ReactDOMServer from 'react-dom/server';
+
+test('navigation-next server side rendering', async () => {
+  (await getExamplesFor('navigation-next')).forEach(examples => {
+    expect(
+      () =>
+        // $StringLitteral
+        ReactDOMServer.renderToString(require(examples.filePath).default), // eslint-disable-line global-require, import/no-dynamic-require
+    ).not.toThrowError();
+  });
+});